Verify Cursor::move_forward (#183)
* Add "lower half" for pte_index_add_with_carry

The new implementation takes care of the cases when cur_level < add_level.

* Add pte_index sequence to int and prove its injectivity

(over sequences with the same length)

* Outline of new proof for move_forward

* Prove new version of lemma_aligned_pte_index_unchanged

* Rewrite transitivity proof using calc!

* Prove carry ends at nonzero result

* Alternative characterization of pte_index

* Prove second part of alt spec

* Remove old proof fn and prove addr align lemma

* Two more properties of page_size

* Relax requirements of carry_ends_at_nonzero

* Verify move_forward

* Rewrite proof of aligned_pte_index_unchanged

using the alternative spec for pte_index.

I think this proof is easier to understand than the one with a ton of bit operations.

* Remove unused lemma

* Format code and remove unused imports

* Add spinoff_prover to take_next

Perhaps this can fix the flaky proof problem?